# Create an initial set of benchmarks and make the first improvements. Custom check for checksum equality.