--- a/go.mod
+++ b/go.mod
@@ -1,0 +1,3 @@
+module immcheck
+
+go 1.15