--- a/go.mod
+++ b/go.mod
@@ -1,3 +1,8 @@
 module immcheck
 
 go 1.15
+
+require (
+	github.com/davecgh/go-spew v1.1.1
+	github.com/sergi/go-diff v1.2.0
+)