… .gitignore
… .golangci.json
… go.sum
… immcheck_test.go
… new.txt
… old.txt
… results.txt