Results:

```
name                                                         old time/op    new time/op    delta
ImmcheckBytes/[16384]byte;muts(0%)-16                          2.07ms ± 1%    2.10ms ± 1%   +1.25%  (p=0.000 n=9+10)
ImmcheckBytes/[16384]byte;muts(1%)-16                          2.11ms ± 1%    2.11ms ± 1%     ~     (p=0.436 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoOrig-16                   2.08ms ± 1%    2.09ms ± 0%   +0.66%  (p=0.028 n=10+9)
ImmcheckBytes/[16384]byte;muts(1%);NoOrig-16                   2.10ms ± 1%    2.10ms ± 1%     ~     (p=0.631 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoSnap-16                   8.98µs ± 1%    8.41µs ± 2%   -6.34%  (p=0.000 n=8+9)
ImmcheckBytes/[16384]byte;muts(1%);NoSnap-16                   9.38µs ± 1%    8.83µs ± 1%   -5.93%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoSnap;NoOrig-16            7.09µs ± 1%    6.57µs ± 1%   -7.35%  (p=0.000 n=9+9)
ImmcheckBytes/[16384]byte;muts(1%);NoSnap;NoOrig-16            7.48µs ± 0%    6.98µs ± 1%   -6.59%  (p=0.000 n=8+9)
ImmcheckTransactions/[8]txs(1024);muts(0%)-16                  11.0ms ± 1%     9.9ms ± 1%  -10.19%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%)-16                  11.3ms ± 5%    10.6ms ± 5%   -6.30%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoOrig-16           10.9ms ± 1%     9.9ms ± 1%   -9.04%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoOrig-16           11.4ms ± 7%    10.6ms ± 7%   -7.11%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoSnap-16           10.9ms ± 1%     9.9ms ± 1%   -9.15%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoSnap-16           11.7ms ± 9%    10.5ms ± 5%  -10.38%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoSnap;NoOrig-16    10.9ms ± 1%     9.9ms ± 1%   -9.40%  (p=0.000 n=10+9)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoSnap;NoOrig-16    11.7ms ± 6%    10.4ms ± 1%  -11.28%  (p=0.000 n=10+9)

name                                                         old alloc/op   new alloc/op   delta
ImmcheckBytes/[16384]byte;muts(0%)-16                          1.00MB ± 0%    1.00MB ± 0%   -0.02%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%)-16                          1.03MB ± 2%    1.03MB ± 2%     ~     (p=0.684 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoOrig-16                   1.00MB ± 0%    1.00MB ± 0%   -0.02%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%);NoOrig-16                   1.03MB ± 2%    1.02MB ± 2%     ~     (p=0.436 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoSnap-16                   1.30kB ± 0%    1.09kB ± 0%  -16.05%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%);NoSnap-16                   1.49kB ± 0%    1.28kB ± 0%  -13.90%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoSnap;NoOrig-16              864B ± 0%      656B ± 0%  -24.07%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%);NoSnap;NoOrig-16            1.06kB ± 0%    0.85kB ± 1%  -19.67%  (p=0.000 n=9+9)
ImmcheckTransactions/[8]txs(1024);muts(0%)-16                  2.46MB ± 0%    1.98MB ± 0%  -19.20%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%)-16                  2.63MB ±10%    2.31MB ±14%  -12.03%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoOrig-16           2.46MB ± 0%    1.98MB ± 0%  -19.22%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoOrig-16           2.70MB ±16%    2.33MB ±14%  -13.70%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoSnap-16           2.46MB ± 0%    1.98MB ± 0%  -19.21%  (p=0.000 n=8+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoSnap-16           2.76MB ±19%    2.26MB ±12%  -18.18%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoSnap;NoOrig-16    2.46MB ± 0%    1.98MB ± 0%  -19.22%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoSnap;NoOrig-16    2.74MB ±13%    2.19MB ± 1%  -20.06%  (p=0.000 n=10+9)

name                                                         old allocs/op  new allocs/op  delta
ImmcheckBytes/[16384]byte;muts(0%)-16                           32.8k ± 0%     32.8k ± 0%   -0.02%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%)-16                           32.8k ± 0%     32.8k ± 0%   -0.02%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoOrig-16                    32.8k ± 0%     32.8k ± 0%   -0.02%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%);NoOrig-16                    32.8k ± 0%     32.8k ± 0%   -0.03%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoSnap-16                     20.0 ± 0%      12.0 ± 0%  -40.00%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%);NoSnap-16                     23.0 ± 0%      15.0 ± 0%  -34.78%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(0%);NoSnap;NoOrig-16              16.0 ± 0%       8.0 ± 0%  -50.00%  (p=0.000 n=10+10)
ImmcheckBytes/[16384]byte;muts(1%);NoSnap;NoOrig-16              19.0 ± 0%      11.0 ± 0%  -42.11%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%)-16                   26.1k ± 0%      1.1k ± 0%  -95.80%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%)-16                   27.4k ± 7%      3.6k ±68%  -86.75%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoOrig-16            26.1k ± 0%      1.1k ± 0%  -95.82%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoOrig-16            27.9k ±11%      3.7k ±67%  -86.66%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoSnap-16            26.1k ± 0%      1.1k ± 0%  -95.81%  (p=0.000 n=9+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoSnap-16            28.4k ±14%      3.2k ±66%  -88.70%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(0%);NoSnap;NoOrig-16     26.1k ± 0%      1.1k ± 0%  -95.83%  (p=0.000 n=10+10)
ImmcheckTransactions/[8]txs(1024);muts(1%);NoSnap;NoOrig-16     28.2k ±10%      2.7k ± 7%  -90.55%  (p=0.000 n=10+9)
```